+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `Client.iter_commands`, `Client.iter_message_commands` and `Client.iter_slash_commands`.
+- Ephemeral default is now applicable at a client and component level with it defaulting to `None` on
+  components, this will propagate down from the client to the command being executed with each level
+  having the option to override its state or leave it as is.
+
+### Changed
+- SlashCommand's ephemeral default now defaults to `None` indicating that the parent entity's state should
+  be used.
 
 ### Removed
 - `tanjun.abc.ExecutableCommand.execute` and `check_context` as this doesn't work typing wise.